+++ pkg/backup/backupDeleteManager.go
@@ -3,6 +3,7 @@ package backup
 import (
 	"context"
 	"fmt"
+
 	k8sv1 "github.com/cloudogu/k8s-backup-lib/api/v1"
 	"github.com/cloudogu/k8s-backup-operator/pkg/provider"
 )
@@ -40,7 +41,7 @@ func (bdm *backupDeleteManager) delete(ctx context.Context, backup *k8sv1.Backup
 }
 
 func (bdm *backupDeleteManager) triggerBackupDelete(ctx context.Context, backup *k8sv1.Backup) error {
-	backupProvider, err := provider.Get(ctx, backup, backup.Spec.Provider, backup.Namespace, bdm.recorder, bdm.k8sClient, bdm.clientSet)
+	backupProvider, err := provider.Get(ctx, backup, backup.Spec.Provider, backup.Namespace, bdm.recorder, bdm.k8sClient)
 	if err != nil {
 		return fmt.Errorf("failed to get backup provider: %w", err)
 	}