+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- [#79] Only cleanup dogus before restoring
+- [#80] Backups now only contain the following:
+  - Dogus, Configmaps, Secrets and PVCs
+    - **This is a breaking change in the sense of how backups in general work.**
 
 ## [v1.6.0] - 2025-09-03
 ### Changed